+++ lib/alf-shell/alf/shell/command/show.rb
@@ -33,14 +33,16 @@ module Alf
         requester.rendering_options[:float_format] = @ff unless @ff.nil?
         requester.renderer_class = (@renderer_class || requester.renderer_class)
 
+        database = requester && requester.database
+
         # normalize args
         args = argv.first
         args = [ stdin_reader ] if args.empty?
-        chain = Iterator.coerce(args.first, requester && requester.database)
+        chain = Iterator.coerce(args.first, database)
 
         # put a sorter
         if argv[1]
-          chain = Alf::Operator::NonRelational::Sort.new([chain], argv[1])
+          chain = Alf::Operator::NonRelational::Sort.new(database, [chain], argv[1])
         end
 
         chain
@@ -48,13 +50,13 @@ module Alf
     
       private 
 
-      def stdin_reader
-        if requester && requester.respond_to?(:stdin_reader)
-          requester.stdin_reader
-        else 
-          Reader.coerce($stdin)
+        def stdin_reader
+          if requester && requester.respond_to?(:stdin_reader)
+            requester.stdin_reader
+          else
+            Reader.coerce($stdin)
+          end
         end
-      end
 
     end # class Show
   end # module Shell