+++ spec/unit/alf-core/tools/test_to_lispy.rb
@@ -121,23 +121,26 @@ module Alf
 
     end # Summarization
 
+    let(:lispy){ examples_database.send(:lispy) }
+
+
     describe "on an nullary operator" do
-      let(:value){ Alf::Operator::NonRelational::Generator.new([], 10, :id) } 
+      let(:value){ lispy.generator(10, :id) } 
       it { should eq("(generator 10, :id)") }
     end
 
     describe "on an monadic operator with an option" do
-      let(:value){ Alf::Operator::Relational::Project.new([:suppliers], [:city], :allbut => true) } 
+      let(:value){ lispy.project(:suppliers, [:city], :allbut => true) } 
       it { should eq("(project :suppliers, [:city], {:allbut => true})") }
     end
 
     describe "on an monadic operator with default values for options" do
-      let(:value){ Alf::Operator::Relational::Project.new([:suppliers], [:city]) } 
+      let(:value){ lispy.project(:suppliers, [:city]) } 
       it { should eq("(project :suppliers, [:city])") }
     end
 
     describe "on an dyadic operator without no args nor any option" do
-      let(:value){ Alf::Operator::Relational::Join.new([:suppliers, :cities]) } 
+      let(:value){ lispy.join(:suppliers, :cities) } 
       it { should eq("(join :suppliers, :cities)") }
     end